Refactor TimeReport to use Formatter module for table formatting - Extracted table formatting logic into a new `Formatter` module. - Removed redundant methods (`format_table_header`, `format_table_separator`, `format_table_row`, `format_notes`) from the `TimeReport` class. - Included the `Formatter` module in the `TimeReport` class to utilize its formatted output.
